requirements.txt
Babel==2.5.1
boto3==1.4.8
botocore==1.8.3
dj-database-url==0.4.2
Django==1.11.5
django-hashid-field==1.2.3
django-phonenumber-field==1.3.0
django-storages==1.6.5
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
hashids==1.2.0
jmespath==0.9.3
olefile==0.44
phonenumberslite==8.8.7
Pillow==4.2.1
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
s3transfer==0.1.11
six==1.11.0
reportlab==3.5.67